Add pipe with three functions
So that it would be easier for consumers to pipe values through more
functions.

I accomplished this by using TypeScript function overloading to keep the
type information even though the function has to be implemented
generically using unknown and Function types.

[changelog]
added: support to pipe a value into a single function
added: support to pipe a value through three functions

ps-id: A116D1FE-436B-4B16-B5E1-EC59F0F98A83