Add compose of three functions
So that consumers would be able to easily compose three functions.

I accomplished this by using TypeScript function overloading to keep the
typing information even though the generic implementation is implemented
with unknowns. It is sad that TypeScript is limited in such a fashion
and doesn't support true function overloads. But this is what we have to
work with so might as well trying and make the best of it.

[changelog]
added: support for composing three functions

ps-id: 4174270B-ED08-43A2-BA22-6373524915F5